Face, Video scrape, Audio scrape, subtitles
This is the current Facedetection3.0 code which implements the video scraping into the OpenCV display. the audio scraping is used in unison with this to display the subtitles as well. And the whisper ai setup is included in this as well. All three of these work together to set up the face detection and text to speech all in one environment with subtitles.